requirements.txt
pandas
requests
uvicorn
python-dotenv
pydub
webrtcvad-wheels
resemblyzer
sqlalchemy
psycopg2-binary
pgvector
flask
flask[async]